Fix bug in FordFulkerson.hpp

+++ include/FordFulkerson.hpp
@@ -50,7 +50,7 @@ void FordFulkerson(GT& G, umap_WT& c, T s, T t, umap_WT& f) {
 		}
 		umap<T, BFSInfo<T>> BFS_ans;
 		BFS(Gf, s, BFS_ans);
-		std::vector<size_t> p;
+		std::vector<int> p;
 		PrintPath(s, t, BFS_ans, p);
 		if (!p.size())
 			break;
